requirements.txt
asgiref==3.11.0
attrs==25.4.0
distlib==0.4.0
Django==6.0.1
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
filelock==3.24.3
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mysql-connector-python==9.5.0
packaging==26.0
platformdirs==4.9.2
PyMySQL==1.1.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
setuptools==80.10.2
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
virtualenv==20.38.0
virtualenvwrapper-win==1.2.7
wheel==0.46.3
